Handle paths with odd characters that cause shell execs to fail.
For example, a space. The current code creates the mountpoint just fine,
but when mounting, the exec call falls over. Escaping or pre-quoting the path
given to mountutils.mount() etc does not help because fs.mkdirSync() makes a
directory that includes the escapes or quotes. The shell consumes the escape
characters, resulting in "directory not found".
To fix, apply correct quoting to paths being passed to the shell exec.